+++ app/jobs/app_version_checking_job.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class AppVersionCheckingJob < ApplicationJob
-  queue_as :default
+  queue_as :app_version_checking
   sidekiq_options retry: false
 
   def perform