+++ app/jobs/cache/preheating_job.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Cache::PreheatingJob < ApplicationJob
-  queue_as :default
+  queue_as :cache
 
   def perform
     User.find_each do |user|