+++ db/migrate/20250728191359_add_country_name_to_points.rb
@@ -1,0 +1,10 @@
+class AddCountryNameToPoints < ActiveRecord::Migration[8.0]
+  disable_ddl_transaction!
+
+  def change
+    add_column :points, :country_name, :string
+    add_index :points, :country_name, algorithm: :concurrently
+
+    DataMigrations::BackfillCountryNameJob.perform_later
+  end
+end